Ran into some happy accidents :D
So my reset button somehow turned into a pause button and then by trying to fix it I somehow ended up with an extra timer replay button. Also i set the timer default to 15 minutes but this can be changed by changing the 15 in the box to anything else, but once the you start the timer and try to change it again it gets funky and you have to reload the page to set new time. I'll probably continue working on it tommorrow.