fixed pause and refresh glitch
If you were to pause the interval and then leave the page, then interval would launch with a different time as if you had left it running but it was paused the whole time!

Fixed it with an if else statement and added little pause equals to true/false assignments for everytime the pause or resume buttons were pressed.