fixed pause and refresh glitch (again)....
Was having problems because for some reason an if statement was still running even though the provided paremeter was false.
 
It turns out if you save a boolean value in the localStorage it converts it into a string so you have to convert it back into a bool value when when reusing it (the same goes for dates, and i think int or float values).