Timer now remains paused on page reload if it was already paused, and updated home page.
The timer used to run wild without the user consent if it was paused on a previous session. Even though the time was accurate if i had the timer paused before then  i want it paused when i come back Damit!

Fixed it by having an if statement clear the interval if the variable "shouldBePaused" is equal to "true".